'Added map'

+++ app/src/main/java/com/example/proyectodamn/local/AuthLocal.java
@@ -38,9 +38,9 @@ public class AuthLocal implements Serializable {
     @AllArgsConstructor
     public static class Geolocalizacion implements Serializable {
         @SerializedName("longitud")
-        public Integer longitud;
+        public double longitud;
         @SerializedName("latitud")
-        public Integer latitud;
+        public double latitud;
 
     }
